Added a couple more opcodes, fixed some small bugs and added a comment directive to the compiler

+++ symbol.h
@@ -18,7 +18,7 @@ typedef struct _symbol {
     char *name;
     char *signature; // used only when type = SYM_METHOD
     int   type;
-    int   cp_index;
+    unsigned short cp_index;
     union {
         unsigned char byte_val;
         int int_val;
